backend/migrations/models/3_20251012183854_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        DROP TABLE IF EXISTS "chatfile";"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        """
