backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.1",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.118.0",
    "openai>=1.44.0",
    "pinecone>=7.3.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.7.0",
    "tortoise-orm[asyncpg]>=0.25.1",
    "tqdm>=4.66.0",
    "uvicorn>=0.37.0",
    "chromadb>=0.5.0",
]

[tool.aerich]
tortoise_orm = "config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
